restructure repo
Restructuring the repo to support a monolithic architecture. This commit moves the cmd, internal, and internal/pkg directories to the backend directory. A directory for the frontend will be added in a future commit.
